requirements.txt
flask
frozen-flask
markdown
pillow
